api_service/requirements.txt
Django==4.0.4
celery
djangorestframework
psycopg2
ipdb